# legal-contracts-front/src/api/models/LegalContract.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LegalContract = {
    id: number;
    author?: string | null;
    entityName?: string | null;
    description?: string | null;
    createdAt?: string;
    updatedAt?: string | null;
};