# legal-contracts-front/src/api/models/LegalContractCreateDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LegalContractCreateDto = {
    author?: string | null;
    entityName?: string | null;
    description?: string | null;
    createdAt?: string;
    updatedAt?: string | null;
};